+++ src/app/app-routing.module.ts
@@ -1,4 +1,4 @@
-import { NgModule } from '@angular/core';
+import { NgModule, Component } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { AboutComponent } from './about/about.component';
 import { HomeComponent } from './home/home.component';
@@ -14,6 +14,8 @@ import { SignupPersonalComponent } from './auth/signup/forms/signup-personal.com
 import { SignupPhoneComponent } from './auth/signup/forms/signup-phone.component';
 import { SignupSailingComponent } from './auth/signup/forms/signup-sailing.component';
 import { VolunteerComponent } from './volunteer/volunteer.component'
+import { ReservationComponent } from './reservation/reservation.component';
+
 
 const routes: Routes = [
 	{ path: '', component: HomeComponent },
@@ -30,6 +32,8 @@ const routes: Routes = [
 	{ path: 'signup/phone', component: SignupPhoneComponent },
 	{ path: 'signup/sailing', component: SignupSailingComponent },
 	{ path: 'volunteer', component: VolunteerComponent }
+	{ path: 'reservation', component: ReservationComponent }
+
 ];
 @NgModule({
 	imports: [RouterModule.forRoot(routes)],